HDDS-11076. Revert HDDS-11076 and HDDS-11078 (#7004)
* Revert "HDDS-11076. NoSuchMethodError: ByteBuffer.position compiling with Java 9+, running with Java 8 (#6897)"

This reverts commit 8f610da91f32ec9e8a124b6cbd211407f0290ca3.

* Revert "HDDS-11078. Remove usage of sun.misc.Signal (#6876)"

This reverts commit 3f1421963922df3101bc2ee82a9c9af5fed7b297.
