+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/hdds/StringUtils.java
@@ -20,9 +20,19 @@ package org.apache.hadoop.hdds;
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Map;
 
 import com.google.common.base.Preconditions;
+import org.apache.commons.lang3.SystemUtils;
+import org.apache.hadoop.hdds.conf.OzoneConfiguration;
+import org.apache.hadoop.hdds.utils.SignalLogger;
+import org.apache.hadoop.hdds.utils.VersionInfo;
+import org.apache.hadoop.net.NetUtils;
+import org.apache.hadoop.util.ShutdownHookManager;
 import org.apache.ratis.thirdparty.io.netty.buffer.Unpooled;
+import org.slf4j.Logger;
 
 /**
  * Simple utility class to collection string conversion methods.
@@ -33,6 +43,11 @@ public final class StringUtils {
   }
 
   private static final Charset UTF8 = StandardCharsets.UTF_8;
+
+  /**
+   * Priority of the StringUtils shutdown hook.
+   */
+  private static final int SHUTDOWN_HOOK_PRIORITY = 0;
 
   /**
    * Decode a specific range of bytes of the given byte array to a string
@@ -90,6 +105,70 @@ public final class StringUtils {
    */
   public static byte[] string2Bytes(String str) {
     return str.getBytes(UTF8);
+  }
+
+  /**
+   * Return a message for logging.
+   * @param prefix prefix keyword for the message
+   * @param msg content of the message
+   * @return a message for logging
+   */
+  public static String toStartupShutdownString(String prefix, String... msg) {
+    StringBuilder b = new StringBuilder(prefix);
+    b.append("\n/************************************************************");
+    for (String s : msg) {
+      b.append("\n").append(prefix).append(s);
+    }
+    b.append("\n************************************************************/");
+    return b.toString();
+  }
+
+  public static void startupShutdownMessage(VersionInfo versionInfo,
+      Class<?> clazz, String[] args, Logger log, OzoneConfiguration conf) {
+    final String hostname = NetUtils.getHostname();
+    final String className = clazz.getSimpleName();
+
+    if (log.isInfoEnabled()) {
+      log.info(createStartupShutdownMessage(versionInfo, className, hostname,
+          args, HddsUtils.processForLogging(conf)));
+    }
+
+    if (SystemUtils.IS_OS_UNIX) {
+      try {
+        SignalLogger.INSTANCE.register(log);
+      } catch (Throwable t) {
+        log.warn("failed to register any UNIX signal loggers: ", t);
+      }
+    }
+    ShutdownHookManager.get().addShutdownHook(
+        () -> log.info(toStartupShutdownString("SHUTDOWN_MSG: ",
+            "Shutting down " + className + " at " + hostname)),
+        SHUTDOWN_HOOK_PRIORITY);
+
+  }
+
+  /**
+   * Generate the text for the startup/shutdown message of processes.
+   * @param className short name of the class
+   * @param hostname hostname
+   * @param args Command arguments
+   * @return a string to log.
+   */
+  public static String createStartupShutdownMessage(VersionInfo versionInfo,
+      String className, String hostname, String[] args,
+      Map<String, String> conf) {
+    return toStartupShutdownString("STARTUP_MSG: ",
+        "Starting " + className,
+        "  host = " + hostname,
+        "  args = " + (args != null ? Arrays.asList(args) : new ArrayList<>()),
+        "  version = " + versionInfo.getVersion(),
+        "  classpath = " + System.getProperty("java.class.path"),
+        "  build = " + versionInfo.getUrl() + "/"
+            + versionInfo.getRevision()
+            + " ; compiled by '" + versionInfo.getUser()
+            + "' on " + versionInfo.getDate(),
+        "  java = " + System.getProperty("java.version"),
+        "  conf = " + conf);
   }
 
   public static String appendIfNotPresent(String str, char c) {
