+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/hdds/utils/SignalLogger.java
@@ -18,17 +18,9 @@
 
 package org.apache.hadoop.hdds.utils;
 
-import jnr.constants.platform.Signal;
-import jnr.posix.POSIX;
-import jnr.posix.POSIXFactory;
-import jnr.posix.SignalHandler;
-import jnr.posix.util.DefaultPOSIXHandler;
 import org.apache.hadoop.hdds.annotation.InterfaceAudience;
 import org.apache.hadoop.hdds.annotation.InterfaceStability;
 import org.slf4j.Logger;
-
-import java.util.EnumSet;
-import java.util.Set;
 
 /**
  * This class logs a message whenever we're about to exit on a UNIX signal.
@@ -42,18 +34,20 @@ public enum SignalLogger {
 
   INSTANCE;
 
-  private static final Set<Signal> SIGNALS = EnumSet.of(Signal.SIGHUP, Signal.SIGINT, Signal.SIGTERM);
+  private static final String[] SIGNALS = new String[] {"TERM", "HUP", "INT"};
 
   private boolean registered = false;
 
   /**
    * Our signal handler.
    */
-  private static class Handler implements SignalHandler {
+  private static class Handler implements sun.misc.SignalHandler {
     private final Logger log;
+    private final sun.misc.SignalHandler prevHandler;
 
-    Handler(Logger log) {
+    Handler(String name, Logger log) {
       this.log = log;
+      prevHandler = sun.misc.Signal.handle(new sun.misc.Signal(name), this);
     }
 
     /**
@@ -62,8 +56,10 @@ public enum SignalLogger {
      * @param signal The incoming signal
      */
     @Override
-    public void handle(int signal) {
-      log.error("RECEIVED SIGNAL {}: {}", signal, Signal.valueOf(signal));
+    public void handle(sun.misc.Signal signal) {
+      log.error("RECEIVED SIGNAL {}: SIG{}",
+          signal.getNumber(), signal.getName());
+      prevHandler.handle(signal);
     }
   }
 
@@ -80,12 +76,10 @@ public enum SignalLogger {
     StringBuilder bld = new StringBuilder();
     bld.append("registered UNIX signal handlers for [");
     String separator = "";
-    final POSIX posix = POSIXFactory.getPOSIX(new DefaultPOSIXHandler(), true);
-    final Handler handler = new Handler(log);
-    for (Signal signal : SIGNALS) {
+    for (String signalName : SIGNALS) {
       try {
-        posix.signal(signal, handler);
-        bld.append(separator).append(signal.name());
+        new Handler(signalName, log);
+        bld.append(separator).append(signalName);
         separator = ", ";
       } catch (Exception e) {
         log.debug("", e);
